third_party/WebKit/LayoutTests/editing/text-iterator/basic-iteration-shadowdom-expected.txt

CONSOLE WARNING: line 22: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Unit tests for WebCore text iterator with shadow tree support enabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS addShadowTreeWithDivElement(div); range.selectNodeContents(div.shadowRoot); internals.rangeAsText(range) is "b"
PASS appendBrElement(div.shadowRoot.childNodes[0]); range.selectNodeContents(div.shadowRoot); internals.rangeAsText(range) is "b\n"
PASS successfullyParsed is true

TEST COMPLETE
